Use generated GDBus for libgdmgreeter
GdmGreeterClient was just a wrapper over the DBus interface, and
was in fact very similar to the autogenerated bindings. Bite the
bullet and replace GdmGreeterClient with generated GdmGreeterServer.
Also add two convenience functions for creating GdmGreeterServers,
one for actual greeters and one for session components that want
to use the greeter interface.

https://bugzilla.gnome.org/show_bug.cgi?id=676381